Pull the api call for the data out of the timeline component and into the homepage so the data is available to the stats page as well without having to make a second data fetch
