+++ src/components/HomePage/HomePage.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 
 import './HomePage.scss';
 import TimeLine from '../TimeLine/TimeLine';
@@ -6,11 +6,13 @@ import Stats from '../Stats/Stats';
 import LaunchDetails from '../LaunchDetails/LaunchDetails';
 import Logo from '../Logo/Logo';
 import Nav from '../Nav/Nav';
+import apiService from '../../utilities/apiService';
 
 export default function HomePage (props) {
 
     const [activeLaunch, setActiveLaunch] = useState(null);
     const [currentNav, setCurrentNav] = useState('timeline');
+    const [launchData, setLaunchData] = useState([]);
 
     function updateActiveLaunch (launch) {
         setActiveLaunch(launch);
@@ -20,12 +22,26 @@ export default function HomePage (props) {
         setCurrentNav(page);
     }
 
+    async function fetchData () {
+        try {
+            const apiData = await apiService.getData();
+            setLaunchData(apiData);
+        } catch (error) {
+            console.log('Unable to retreive data');
+        }
+    }
+
+    useEffect(() => {
+        fetchData();
+    }, [])
+
     return (
         <div id="homepage-container">
             <img src={'/rocket.jpg'} alt="SpaceX rocket taking off" />
             {
-                currentNav === 'timeline' ? 
-                    <TimeLine activeLaunch={activeLaunch} updateActiveLaunch={updateActiveLaunch} />
+                currentNav === 'timeline' ?
+                    launchData.length ? <TimeLine launchData={launchData} activeLaunch={activeLaunch} updateActiveLaunch={updateActiveLaunch} />
+                    : ''
                 : <Stats />
             }
             <Logo />
