+++ src/components/Stats/Stats.jsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
 import './Stats.scss';
+import CustomerPieChart from '../CustomerPieChart/CustomerPieChart';
 
-export default function Stats (props) {
+export default function Stats ({ launches }) {
     return (
         <div id="stats-container">
             Stats to be made next...
